feat: implement link management in Toolbar with LinkPopup component
